Constants changed
I made changes of the conversion constants for position and velocity sets. Also I found a mistake which would perhaps cause a problem that was noticed during testing. Line 183 had a == to number of points, which would only enable the talon had only that amount of points. Essential a small burst of motion. Now it is set when greater than. Needs to be tested.